r/aws_lambda_function: Add 'skip_destroy' argument.
Acceptance test output:

% make testacc TESTARGS='-run=TestAccLambdaFunction_basic\|TestAccLambdaFunction_skipDestroy\|TestAccLambdaFunction_disappears' PKG=lambda ACCTEST_PARALLELISM=3
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go test ./internal/service/lambda/... -v -count 1 -parallel 3  -run=TestAccLambdaFunction_basic\|TestAccLambdaFunction_skipDestroy\|TestAccLambdaFunction_disappears -timeout 180m
=== RUN   TestAccLambdaFunction_basic
=== PAUSE TestAccLambdaFunction_basic
=== RUN   TestAccLambdaFunction_disappears
=== PAUSE TestAccLambdaFunction_disappears
=== RUN   TestAccLambdaFunction_skipDestroy
=== PAUSE TestAccLambdaFunction_skipDestroy
=== CONT  TestAccLambdaFunction_basic
=== CONT  TestAccLambdaFunction_skipDestroy
=== CONT  TestAccLambdaFunction_disappears
--- PASS: TestAccLambdaFunction_disappears (37.24s)
--- PASS: TestAccLambdaFunction_basic (45.85s)
--- PASS: TestAccLambdaFunction_skipDestroy (48.66s)
PASS
ok  	github.com/hashicorp/terraform-provider-aws/internal/service/lambda	54.571s

+++ internal/service/lambda/function.go
@@ -291,6 +291,11 @@ func ResourceFunction() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+			"skip_destroy": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  false,
+			},
 			"snap_start": {
 				Type:     schema.TypeList,
 				Optional: true,
@@ -996,6 +1001,11 @@ func resourceFunctionDelete(ctx context.Context, d *schema.ResourceData, meta in
 	var diags diag.Diagnostics
 	conn := meta.(*conns.AWSClient).LambdaClient()
 
+	if v, ok := d.GetOk("skip_destroy"); ok && v.(bool) {
+		log.Printf("[DEBUG] Retaining Lambda Function: %s", d.Id())
+		return diags
+	}
+
 	log.Printf("[INFO] Deleting Lambda Function: %s", d.Id())
 	_, err := tfresource.RetryWhenIsAErrorMessageContains[*types.InvalidParameterValueException](ctx, d.Timeout(schema.TimeoutDelete), func() (interface{}, error) {
 		return conn.DeleteFunction(ctx, &lambda.DeleteFunctionInput{
@@ -1030,14 +1040,15 @@ func FindFunctionByName(ctx context.Context, conn *lambda.Client, name string) (
 
 func findFunction(ctx context.Context, conn *lambda.Client, input *lambda.GetFunctionInput) (*lambda.GetFunctionOutput, error) {
 	output, err := conn.GetFunction(ctx, input)
-	if err != nil {
-		var nfe *types.ResourceNotFoundException
-		if errors.As(err, &nfe) {
-			return nil, &resource.NotFoundError{
-				LastError:   err,
-				LastRequest: input,
-			}
+
+	if errs.IsA[*types.ResourceNotFoundException](err) {
+		return nil, &resource.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
 		}
+	}
+
+	if err != nil {
 		return nil, err
 	}
